Add attribute type to AttributeCreateNode
Changed that the attribute type is required when creating custom attributes.
This ensures that we create the attribute type that the user expects.
Previously it was difficult to differentiate between int and float.
Now the attribute is created regardless of the value, which allows
to create custom attributes without a value, but also to set the value
through references directly when creating the attribute.

The new syntax to create custom attributes:
+<type> <attribute_name>
+<type> <attribute_name> = <value>

The following attribute types are implemented:
- int
- float
- bool
- string
- vec2
- vec4